% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticbuilder-package.R
\docType{package}
\name{elasticbuilder}
\alias{elasticbuilder}
\title{elasticbuilder: Helper to Elasticsearch queries.}
\description{
This package is an helper to build your Elasticsearch queries.
}
\details{
More details will come !
}
